creeated a userAuthMiddleware for authorisation:
