+++ src/routes/userRoute.js
@@ -3,40 +3,22 @@ import { responseClient } from "../middleware/responseClient.js";
 import { verifyAccessJWT } from "../utils/jwt.js";
 import { getSession } from "../models/session/SessionModel.js";
 import { getUserByEmail } from "../models/user/UserModel.js";
+import { userAuthMiddleware } from "../middleware/authMiddleware.js";
 
 const router = express.Router();
 
-router.get("/profile", async (req, res, payload) => {
-  //get accessJWT
-  const { authorization } = req.headers;
-  let message = "unauthorised";
-  if (authorization) {
-    const token = authorization.split(" ")[1];
+router.get("/profile", userAuthMiddleware, async (req, res) => {
+  const user = req.userInfo;
+  user.passsword = undefined
+  user.__v = undefined
+  user.refreshJWT = undefined
 
-    //check if valid
-
-    const decoded = verifyAccessJWT(token);
-
-    //check if the user exist in the session table
-    if (decoded.email) {
-      const tokenSession = await getSession({ token });
-      if (tokenSession?._id) {
-        //get user by email
-        const user = await getUserByEmail(decoded.email);
-        if (user?._id && user.status === "active") {
-          //return the user
-          return responseClient({ 
-            req, 
-            res, 
-            message: "User profile", 
-            payload });
-        }
-      }
-    };
-    message = decoded === "jwt expired" ? decoded : "unauthorised"
-  };
-//   const message = decoded === "jwt expired" ? decoded : "unauthorised"
-  responseClient({ req, res, message, statusCode: 401 });
+  return responseClient({
+    req,
+    res,
+    message: "User profile",
+    payload: user,
+  });
 });
 
 export default router;
